Final version
Final, fully (almost) working input classifier based on the simple genetic algorithm given to us as part of the assignment brief. Also included are several .csv files, which were used to derive graphs and general observations. It should be noted that these csv files are not final; the code itself is finished but I will be running the program continually as i write the report to derive data as needs be.